backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# ASGI for PythonAnywhere
asgiref==3.7.2


